% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_scoretable}
\alias{pca_scoretable}
\title{Extract PCA score table}
\usage{
pca_scoretable(x, naxe = 1)
}
\arguments{
\item{x}{glPCA object}

\item{naxe}{number of axis for which you want to extract PCA scores}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Compute variance explained by a few PCA axis
}
\details{
PCA from glPca. It will return a dataframe, including ID and pca.score for the desire number of axis
}
\examples{
## simulate a toy dataset
x <- adegenet::glSim(50,4e3, 50, ploidy=2)
## perform PCA
pca1 <- adegenet::glPca(x, nf=3)
## Extract variance
res <- pca_scoretable(pca1, naxe=3)
head(res)
}
