% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_var}
\alias{pca_var}
\title{Compute variance explained by first PCA axis}
\usage{
pca_var(x, naxe = 10)
}
\arguments{
\item{x}{glPCA object}

\item{naxe}{number of axis for which you want to know explained variance}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Compute variance explained by a few PCA axis.
}
\details{
PCA from glPca from the package Adegenet
}
\examples{
## simulate a toy dataset
x <- adegenet::glSim(50,4e3, 50, ploidy=2)
## perform PCA
pca1 <- adegenet::glPca(x, nf=3)
## Extract variance
res <- pca_var(pca1, naxe=3)
res
}
