% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_varplot}
\alias{pca_varplot}
\title{Plot variance explained by PCA axis}
\usage{
pca_varplot(x, naxe = 10)
}
\arguments{
\item{x}{glPCA object}

\item{naxe}{number of axis for which you want to know explained variance}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Compute variance explained by a few PCA axis
}
\details{
PCA from glPca
}
\examples{
## simulate a toy dataset
x <- adegenet::glSim(50,4e3, 50, ploidy=2)
## perform PCA
pca1 <- adegenet::glPca(x, nf=3)
## Extract variance
pca_varplot(pca1, naxe=3)
}
