% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{str_readq}
\alias{str_readq}
\title{Read structure q files}
\usage{
str_readq(file)
}
\arguments{
\item{file}{Path to the file}
}
\description{
str_readq(.) read structure q file as produce by faststructure
}
\details{
more details to come ...
}
\examples{
# provide some examples of how to use your function
 file.path <- ".results_job_k3_r1_q" 
 str_readq(file.path)

}
