% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{str_readsummary}
\alias{str_readsummary}
\title{Read structure summary file}
\usage{
str_readsummary(file)
}
\arguments{
\item{file}{Path to the file}
}
\description{
str_readsummary(.) read structure  summary file as produced by faststructure
}
\details{
more details to come ...
}
\examples{
# provide some examples of how to use your function
 file.path <- ".results_job_k3_r1_q" 
 str_readsummary(file.path)

}
