% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallele_jobs.R
\name{strc_create_jobs}
\alias{strc_create_jobs}
\title{Create a job file for faststructure package}
\usage{
strc_create_jobs(
  nk = 1:4,
  nrep = 3,
  popnum = 1,
  burn = 1000,
  iteration = 10000
)
}
\arguments{
\item{nk}{Numeric vector of the K values to be tested}

\item{nrep}{Number of repetition at each K values}

\item{popnum}{Numeric vector of the pop that you want to use}

\item{burn}{Number of burnin}

\item{iteration}{Number of iteration}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a job file for faststructure package
}
\details{
more details to come ...
}
\examples{
# create a new job
job1 <- strc_create_jobs(nk = 1:2, nrep = 3, popnum = 1:2, burn = 1000, iteration = 10000)
job1
# write(t(jobs1), ncol=length(jobs1[1,]), file='path/to/file.txt')
}
