% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{strc_readq}
\alias{strc_readq}
\title{Read structure q files}
\usage{
strc_readq(file.path)
}
\arguments{
\item{file.path}{Path to the file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
strc_readq(.) read structure q file as produce by faststructure
}
\details{
more details to come ...
}
\examples{
# provide some examples of how to use your function
path.to.file <- system.file("extdata", "results_job_k2_r1_q", package = "QuickPop")
strc_readq(path.to.file)
}
