% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{strc_readsummary}
\alias{strc_readsummary}
\title{Read structure summary file}
\usage{
strc_readsummary(file.path)
}
\arguments{
\item{file.path}{Path to the file}
}
\description{
strc_readsummary(.) read structure summary file as produced by faststructure
}
\details{
Read directly the file as produced by faststructure package
}
\examples{
# provide some examples of how to use your function
#path.to.file <- "data-raw/results_summary.csv"
path.to.file <- system.file("extdata", "results_summary.csv", package = "QuickPop")
strc_readsummary(path.to.file)
}
