% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strcSummary.data}
\alias{strcSummary.data}
\title{Structure run summary}
\format{
A tibble with 30 rows and 8 variables
}
\usage{
data(strcSummary.data)
}
\description{
test

\itemize{
\item run_id. to be define
\item burnin.
\item iteration.
\item k.
\item ln_prob_data.
\item mean_llh.
\item var_llh.
\item mean_alpha.
}
}
