% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strcSummary}
\alias{strcSummary}
\title{Structure run summary}
\format{
A tibble with 30 rows and 8 variables
}
\usage{
data(strcSummary)
}
\description{
test

\itemize{
\item run_id. price in US dollars
\item brunin. weight of the diamond
}
}
