% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evanno.R
\name{strc_evanno}
\alias{strc_evanno}
\title{Compute Evanno}
\usage{
strc_evanno(df)
}
\arguments{
\item{df}{Dataframe with structure results}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Compute classical Evanno from structure summary file and return a graphic
}
\details{
more details to come ...
}
\examples{
head(strcSummary.data)
strc_evanno(strcSummary.data)

}
