% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evanno.R
\name{strc_evanno}
\alias{strc_evanno}
\title{Compute Evanno}
\usage{
strc_evanno(df)
}
\arguments{
\item{df}{Dataframe with structure results}
}
\description{
Compute classical Evanno from structure summary file and return a graphic
}
\details{
more details to come ...
}
\examples{
# provide some examples of how to use your function
strc_evanno(strcSummary)

}
