% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{strc_readq}
\alias{strc_readq}
\title{Read structure q files}
\usage{
strc_readq(file.path)
}
\arguments{
\item{file.path}{Path to the file}
}
\description{
strc_readq(.) read structure q file as produce by faststructure
}
\details{
more details to come ...
}
\examples{
# provide some examples of how to use your function
file.path <- "data-raw/results_job_k2_r1_q"
strc_readq(file.path)
}
