% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{strc_readsummary}
\alias{strc_readsummary}
\title{Read structure summary file}
\usage{
strc_readsummary(file)
}
\arguments{
\item{file}{Path to the file}
}
\description{
strc_readsummary(.) read structure  summary file as produced by faststructure
}
\details{
more details to come ...
}
\examples{
# provide some examples of how to use your function
file.path <- ".results_job_k3_r1_q"
strc_readsummary(file.path)
}
